+++ FactoryAndBuilder/ContinuedEnemy.java
@@ -94,7 +94,7 @@ public class ContinuedEnemy implements Enemy {
 	@Override
 	public void Shoot(int c_degree) {
 		// METODAS KURIS APSKAICIUOJA KULKOS KAMPA
-		double[] a = { 0, 0 };
+		int[] a = { 0, 0 };
 		Big_Fast_Bullet b = F.createFastBullet(false, getX()+65, getY()+65, a) ;
 		System.out.println("ContinuedEnemy shot " + b.getClass() + " at " + c_degree + " degrees");
 		Bullets.add(b);
@@ -106,7 +106,7 @@ public class ContinuedEnemy implements Enemy {
 		Random random = new Random();
     	int ran = random.nextInt(360) + 1;
 		double radians = ran*Math.PI/180.0;
-		double[] bulletVelocity = { Math.cos(radians) * 7, Math.sin(radians) * 7 };
+		int[] bulletVelocity = { (int) (Math.cos(radians) * 7), (int) (Math.sin(radians) * 7) };
 		Big_Fast_Bullet b = F.createFastBullet(false, getX()+35, getY()+35, bulletVelocity) ;
 		System.out.println("Continued shot random " + b.getClass());
 		Bullets.add(b);
